app/requirements.txt
Django==5.2
djangorestframework==3.14.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
djangorestframework-simplejwt==5.5.1